Move plugin scanning logic out of DelegatingClassLoader
