+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/isolation/DelegatingClassLoader.java
@@ -16,27 +16,11 @@
  */
 package org.apache.kafka.connect.runtime.isolation;
 
-import org.apache.kafka.common.config.provider.ConfigProvider;
-import org.apache.kafka.connect.components.Versioned;
-import org.apache.kafka.connect.connector.policy.ConnectorClientConfigOverridePolicy;
-import org.apache.kafka.connect.rest.ConnectRestExtension;
-import org.apache.kafka.connect.sink.SinkConnector;
-import org.apache.kafka.connect.source.SourceConnector;
-import org.apache.kafka.connect.storage.Converter;
-import org.apache.kafka.connect.storage.HeaderConverter;
-import org.apache.kafka.connect.transforms.Transformation;
-import org.apache.kafka.connect.transforms.predicates.Predicate;
-import org.reflections.Configuration;
-import org.reflections.Reflections;
-import org.reflections.ReflectionsException;
-import org.reflections.scanners.SubTypesScanner;
 import org.reflections.util.ClasspathHelper;
-import org.reflections.util.ConfigurationBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLClassLoader;
@@ -44,21 +28,14 @@ import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
-import java.sql.Driver;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.ServiceConfigurationError;
-import java.util.ServiceLoader;
 import java.util.Set;
 import java.util.SortedMap;
-import java.util.SortedSet;
 import java.util.TreeMap;
-import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
@@ -104,6 +81,33 @@ public class DelegatingClassLoader extends URLClassLoader {
         this(pluginLocations, DelegatingClassLoader.class.getClassLoader());
     }
 
+    public Set<PluginSource> sources() {
+        Set<PluginSource> pluginLoaders = new HashSet<>();
+        for (Path pluginLocation : pluginLocations) {
+
+            try {
+                List<URL> pluginUrls = new ArrayList<>();
+                for (Path path : PluginUtils.pluginUrls(pluginLocation)) {
+                    pluginUrls.add(path.toUri().toURL());
+                }
+                URL[] urls = pluginUrls.toArray(new URL[0]);
+                PluginClassLoader loader = newPluginClassLoader(
+                        pluginLocation.toUri().toURL(),
+                        urls,
+                        this
+                );
+                pluginLoaders.add(new PluginSource(pluginLocation, loader, urls));
+            } catch (InvalidPathException | MalformedURLException e) {
+                log.error("Invalid path in plugin path: {}. Ignoring.", pluginLocation, e);
+            } catch (IOException e) {
+                log.error("Could not get listing for plugin path: {}. Ignoring.", pluginLocation, e);
+            }
+        }
+        URL[] classpathUrls = ClasspathHelper.forJavaClassPath().toArray(new URL[0]);
+        pluginLoaders.add(new PluginSource(null, getParent(), classpathUrls));
+        return pluginLoaders;
+    }
+
     /**
      * Retrieve the PluginClassLoader associated with a plugin class
      * @param name The fully qualified class name of the plugin
@@ -147,229 +151,7 @@ public class DelegatingClassLoader extends URLClassLoader {
         );
     }
 
-    public PluginScanResult initLoaders() {
-        List<PluginScanResult> results = new ArrayList<>();
-        for (Path pluginLocation : pluginLocations) {
-            try {
-                results.add(registerPlugin(pluginLocation));
-            } catch (InvalidPathException | MalformedURLException e) {
-                log.error("Invalid path in plugin path: {}. Ignoring.", pluginLocation, e);
-            } catch (IOException e) {
-                log.error("Could not get listing for plugin path: {}. Ignoring.", pluginLocation, e);
-            }
-        }
-        // Finally add parent/system loader.
-        results.add(scanUrlsAndAddPlugins(
-                getParent(),
-                ClasspathHelper.forJavaClassPath().toArray(new URL[0])
-        ));
-        PluginScanResult scanResult = new PluginScanResult(results);
-        installDiscoveredPlugins(scanResult);
-        return scanResult;
-    }
-
-    private PluginScanResult registerPlugin(Path pluginLocation)
-        throws IOException {
-        log.info("Loading plugin from: {}", pluginLocation);
-        List<URL> pluginUrls = new ArrayList<>();
-        for (Path path : PluginUtils.pluginUrls(pluginLocation)) {
-            pluginUrls.add(path.toUri().toURL());
-        }
-        URL[] urls = pluginUrls.toArray(new URL[0]);
-        if (log.isDebugEnabled()) {
-            log.debug("Loading plugin urls: {}", Arrays.toString(urls));
-        }
-        PluginClassLoader loader = newPluginClassLoader(
-                pluginLocation.toUri().toURL(),
-                urls,
-                this
-        );
-        return scanUrlsAndAddPlugins(loader, urls);
-    }
-
-    private PluginScanResult scanUrlsAndAddPlugins(
-            ClassLoader loader,
-            URL[] urls
-    ) {
-        PluginScanResult plugins = scanPluginPath(loader, urls);
-        log.info("Registered loader: {}", loader);
-        loadJdbcDrivers(loader);
-        return plugins;
-    }
-
-    private void loadJdbcDrivers(final ClassLoader loader) {
-        // Apply here what java.sql.DriverManager does to discover and register classes
-        // implementing the java.sql.Driver interface.
-        AccessController.doPrivileged(
-            (PrivilegedAction<Void>) () -> {
-                ServiceLoader<Driver> loadedDrivers = ServiceLoader.load(
-                        Driver.class,
-                        loader
-                );
-                Iterator<Driver> driversIterator = loadedDrivers.iterator();
-                try {
-                    while (driversIterator.hasNext()) {
-                        Driver driver = driversIterator.next();
-                        log.debug(
-                                "Registered java.sql.Driver: {} to java.sql.DriverManager",
-                                driver
-                        );
-                    }
-                } catch (Throwable t) {
-                    log.debug(
-                            "Ignoring java.sql.Driver classes listed in resources but not"
-                                    + " present in class loader's classpath: ",
-                            t
-                    );
-                }
-                return null;
-            }
-        );
-    }
-
-    private PluginScanResult scanPluginPath(
-            ClassLoader loader,
-            URL[] urls
-    ) {
-        ConfigurationBuilder builder = new ConfigurationBuilder();
-        builder.setClassLoaders(new ClassLoader[]{loader});
-        builder.addUrls(urls);
-        builder.setScanners(new SubTypesScanner());
-        builder.useParallelExecutor();
-        Reflections reflections = new InternalReflections(builder);
-
-        return new PluginScanResult(
-                getPluginDesc(reflections, SinkConnector.class, loader),
-                getPluginDesc(reflections, SourceConnector.class, loader),
-                getPluginDesc(reflections, Converter.class, loader),
-                getPluginDesc(reflections, HeaderConverter.class, loader),
-                getTransformationPluginDesc(loader, reflections),
-                getPredicatePluginDesc(loader, reflections),
-                getServiceLoaderPluginDesc(ConfigProvider.class, loader),
-                getServiceLoaderPluginDesc(ConnectRestExtension.class, loader),
-                getServiceLoaderPluginDesc(ConnectorClientConfigOverridePolicy.class, loader)
-        );
-    }
-
-    @SuppressWarnings({"unchecked"})
-    private SortedSet<PluginDesc<Predicate<?>>> getPredicatePluginDesc(ClassLoader loader, Reflections reflections) {
-        return (SortedSet<PluginDesc<Predicate<?>>>) (SortedSet<?>) getPluginDesc(reflections, Predicate.class, loader);
-    }
-
-    @SuppressWarnings({"unchecked"})
-    private SortedSet<PluginDesc<Transformation<?>>> getTransformationPluginDesc(ClassLoader loader, Reflections reflections) {
-        return (SortedSet<PluginDesc<Transformation<?>>>) (SortedSet<?>) getPluginDesc(reflections, Transformation.class, loader);
-    }
-
-    private <T> SortedSet<PluginDesc<T>> getPluginDesc(
-            Reflections reflections,
-            Class<T> klass,
-            ClassLoader loader
-    ) {
-        Set<Class<? extends T>> plugins;
-        try {
-            plugins = reflections.getSubTypesOf(klass);
-        } catch (ReflectionsException e) {
-            log.debug("Reflections scanner could not find any classes for URLs: " +
-                    reflections.getConfiguration().getUrls(), e);
-            return Collections.emptySortedSet();
-        }
-
-        SortedSet<PluginDesc<T>> result = new TreeSet<>();
-        for (Class<? extends T> pluginKlass : plugins) {
-            if (!PluginUtils.isConcrete(pluginKlass)) {
-                log.debug("Skipping {} as it is not concrete implementation", pluginKlass);
-                continue;
-            }
-            if (pluginKlass.getClassLoader() != loader) {
-                log.debug("{} from other classloader {} is visible from {}, excluding to prevent isolated loading",
-                        pluginKlass.getSimpleName(), pluginKlass.getClassLoader(), loader);
-                continue;
-            }
-            try (LoaderSwap loaderSwap = withClassLoader(loader)) {
-                result.add(pluginDesc(pluginKlass, versionFor(pluginKlass), loader));
-            } catch (ReflectiveOperationException | LinkageError e) {
-                log.error("Failed to discover {}: Unable to instantiate {}{}", klass.getSimpleName(), pluginKlass.getSimpleName(), reflectiveErrorDescription(e), e);
-            }
-        }
-        return result;
-    }
-
-    @SuppressWarnings({"rawtypes", "unchecked"})
-    private <T> PluginDesc<T> pluginDesc(Class<? extends T> plugin, String version, ClassLoader loader) {
-        return new PluginDesc(plugin, version, loader);
-    }
-
-    @SuppressWarnings("unchecked")
-    private <T> SortedSet<PluginDesc<T>> getServiceLoaderPluginDesc(Class<T> klass, ClassLoader loader) {
-        SortedSet<PluginDesc<T>> result = new TreeSet<>();
-        ServiceLoader<T> serviceLoader = ServiceLoader.load(klass, loader);
-        for (Iterator<T> iterator = serviceLoader.iterator(); iterator.hasNext(); ) {
-            try (LoaderSwap loaderSwap = withClassLoader(loader)) {
-                T pluginImpl;
-                try {
-                    pluginImpl = iterator.next();
-                } catch (ServiceConfigurationError t) {
-                    log.error("Failed to discover {}{}", klass.getSimpleName(), reflectiveErrorDescription(t.getCause()), t);
-                    continue;
-                }
-                Class<? extends T> pluginKlass = (Class<? extends T>) pluginImpl.getClass();
-                if (pluginKlass.getClassLoader() != loader) {
-                    log.debug("{} from other classloader {} is visible from {}, excluding to prevent isolated loading",
-                            pluginKlass.getSimpleName(), pluginKlass.getClassLoader(), loader);
-                    continue;
-                }
-                result.add(pluginDesc(pluginKlass, versionFor(pluginImpl), loader));
-            }
-        }
-        return result;
-    }
-
-    private static <T>  String versionFor(T pluginImpl) {
-        try {
-            if (pluginImpl instanceof Versioned) {
-                return ((Versioned) pluginImpl).version();
-            }
-        } catch (Throwable t) {
-            log.error("Failed to get plugin version for " + pluginImpl.getClass(), t);
-        }
-        return UNDEFINED_VERSION;
-    }
-
-    public static <T> String versionFor(Class<? extends T> pluginKlass) throws ReflectiveOperationException {
-        // Unconditionally use the default constructor to create an instance to assert that
-        // the constructor exists and can complete successfully.
-        T pluginImpl = pluginKlass.getDeclaredConstructor().newInstance();
-        return versionFor(pluginImpl);
-    }
-
-    private static String reflectiveErrorDescription(Throwable t) {
-        if (t instanceof NoSuchMethodException) {
-            return ": Plugin class must have a no-args constructor, and cannot be a non-static inner class";
-        } else if (t instanceof SecurityException) {
-            return ": Security settings must allow reflective instantiation of plugin classes";
-        } else if (t instanceof IllegalAccessException) {
-            return ": Plugin class default constructor must be public";
-        } else if (t instanceof ExceptionInInitializerError) {
-            return ": Failed to statically initialize plugin class";
-        } else if (t instanceof InvocationTargetException) {
-            return ": Failed to invoke plugin constructor";
-        } else {
-            return "";
-        }
-    }
-
-    public LoaderSwap withClassLoader(ClassLoader loader) {
-        ClassLoader savedLoader = Plugins.compareAndSwapLoaders(loader);
-        try {
-            return new LoaderSwap(savedLoader);
-        } catch (Throwable t) {
-            Plugins.compareAndSwapLoaders(savedLoader);
-            throw t;
-        }
-    }
-
-    private void installDiscoveredPlugins(PluginScanResult scanResult) {
+    public void installDiscoveredPlugins(PluginScanResult scanResult) {
         pluginLoaders.putAll(computePluginLoaders(scanResult));
         for (String pluginClassName : pluginLoaders.keySet()) {
             log.info("Added plugin '{}'", pluginClassName);
@@ -398,26 +180,5 @@ public class DelegatingClassLoader extends URLClassLoader {
                 pluginLoaders.computeIfAbsent(pluginDesc.className(), k -> new TreeMap<>())
                         .put(pluginDesc, pluginDesc.loader()));
         return pluginLoaders;
-    }
-
-    private static class InternalReflections extends Reflections {
-
-        public InternalReflections(Configuration configuration) {
-            super(configuration);
-        }
-
-        // When Reflections is used for parallel scans, it has a bug where it propagates ReflectionsException
-        // as RuntimeException.  Override the scan behavior to emulate the singled-threaded logic.
-        @Override
-        protected void scan(URL url) {
-            try {
-                super.scan(url);
-            } catch (ReflectionsException e) {
-                Logger log = Reflections.log;
-                if (log != null && log.isWarnEnabled()) {
-                    log.warn("could not create Vfs.Dir from url. ignoring the exception and continuing", e);
-                }
-            }
-        }
     }
 }
